refactor: 미사용 파일 제거 및 폴더명에 PascalCase 적용
