--- apps/app/src/screens/BottomTab/type.ts
@@ -1,7 +1,0 @@
-import { FirstProps } from '@/screens/BottomTab/First/type'
-import { SecondProps } from '@/screens/BottomTab/Second/type'
-
-export type BottomTabRoute = {
-  First: FirstProps
-  Second: SecondProps
-}
