--- apps/app/src/screens/RootNativeStack/Example/type.ts
@@ -1,3 +1,0 @@
-export type ExampleProps = {
-  test: string
-}
