+++ apps/app/src/screens/bottom-tab/Second/type.ts
