+++ apps/app/src/screens/bottom-tab/constants.ts
@@ -2,8 +2,8 @@ import { FunctionComponent } from 'react'
 
 import { BottomTabNavigationOptions } from '@react-navigation/bottom-tabs'
 
-import { First } from '@/screens/BottomTab/First'
-import { Second } from '@/screens/BottomTab/Second'
+import { First } from '@/screens/bottom-tab/First'
+import { Second } from '@/screens/bottom-tab/Second'
 
 type BottomTabParams = {
   name: string
