+++ apps/app/src/screens/bottom-tab/type.ts
@@ -1,0 +1,7 @@
+import { FirstProps } from '@/screens/bottom-tab/First/type'
+import { SecondProps } from '@/screens/bottom-tab/Second/type'
+
+export type BottomTabRoute = {
+  First: FirstProps
+  Second: SecondProps
+}
