+++ apps/app/src/screens/root-native-stack/accept-invite/index.tsx
