+++ apps/app/src/screens/root-native-stack/type.ts
@@ -1,6 +1,6 @@
 import { NavigatorScreenParams } from '@react-navigation/native'
 
-import { BottomTabRoute } from '@/screens/BottomTab/type'
+import { BottomTabRoute } from '@/screens/bottom-tab/type'
 
 export type NativeStackList = {
   BottomTab: NavigatorScreenParams<BottomTabRoute>
