+++ apps/app/src/screens/type.ts
@@ -1,3 +1,3 @@
-import { NativeStackList } from '@/screens/RootNativeStack/type'
+import { NativeStackList } from '@/screens/root-native-stack/type'
 
 export type AllScreenType = NativeStackList
